Add files via upload
Joining tables of the feature classes, with a common field to join on. Add a new field to the classes, merge nine feature classes into new class, intersect the merged features with the watershed boundaries. Found the number of features in the intersected feature class that correspond with each watershed, with the 2 numbers reported at the end.